tokenSecurity

+++ src/main/java/com/intouncommon/backend/Controller/commonMethodController.java
@@ -31,13 +31,22 @@ public class commonMethodController {
     }
 
     @PostMapping("/addAdmin")
-    private String addAdmin(@RequestBody admin admin){
-        adminService.addAdmin(admin);
+    private String addAdmin(@RequestBody admin admin , @RequestHeader String header) throws Exception{
+        
+        if(adminService.checkTokenValidity(header)){
+adminService.addAdmin(admin);
         return "added successfully";
+        }
+        return "Wrong token";
+        
     }
     @PutMapping ("/changeAdmin")
-    private String changeAdmin(@RequestBody oldadmin admin){
+    private String changeAdmin(@RequestBody oldadmin admin , @RequestHeader String header) throws Exception{
+        
+        if(adminService.checkTokenValidity(header)){
         return adminService.changeAdmin(admin.getOldUsername(),admin.getUsername(),admin.getOldPassword(),admin.getPassword());
+        }
+        return "Wrong token";
     }
     @GetMapping("/test")
     private String test(@RequestParam String name,@RequestParam String age){
@@ -50,6 +59,7 @@ public class commonMethodController {
     
     @GetMapping("/getvalidity")
     private String validityCheck(@RequestHeader String header) throws Exception {
+        
 
 //        response response = new response();
 //        response.setResponse("test pass");
@@ -67,25 +77,43 @@ public class commonMethodController {
     }
 
     @PostMapping(value="/product/add" ,consumes = "application/json", produces = "application/json")
-    private Long addProduct(@RequestBody productionDto production){
-        return commonMethodService.addProduction(production);
+    private Long addProduct(@RequestBody productionDto production , @RequestHeader String header) throws Exception{
+        
+        if(adminService.checkTokenValidity(header)){
+            return commonMethodService.addProduction(production);
+        }
+        return null;
+        
        // return "Successfully Added";
     }
 
     @PutMapping("/product/update")
-    private String updateProduct(@RequestParam Long id, @RequestBody productions production){
-        return commonMethodService.updateProduction(id,production);
+    private String updateProduct(@RequestParam Long id, @RequestBody productions production,@RequestHeader String header) throws Exception{
+        
+        if(adminService.checkTokenValidity(header)){
+            return commonMethodService.updateProduction(id,production);
+        }
+        return "Wrong token";
     }
 
     @DeleteMapping("/product/delete")
-    private String deleteProduct(@RequestParam Long id){
-        return commonMethodService.deleteProduction(id);
+    private String deleteProduct(@RequestParam Long id , @RequestHeader String header) throws Exception{
+        
+        if(adminService.checkTokenValidity(header)){
+             return commonMethodService.deleteProduction(id);
+        }
+        return "Wrong token";
+       
     }
 
     @PostMapping("/product/url/add")
-    private String addUrl(@RequestBody productImageDTO productImageDTO){
-        commonMethodService.addImageUrl(productImageDTO);
+    private String addUrl(@RequestBody productImageDTO productImageDTO ,@RequestHeader String header) throws Exception{
+        if(adminService.checkTokenValidity(header)){
+             commonMethodService.addImageUrl(productImageDTO);
         return "added";
+        }
+        return "Wrong token";
+       
     }
     /** Category **/
 
@@ -96,81 +124,136 @@ public class commonMethodController {
     }
 
     @PostMapping("/category/add")
-    private String addCategory(@RequestBody categories category){
-        commonMethodService.addCategory(category);
+    private String addCategory(@RequestBody categories category ,@RequestHeader String header) throws Exception{
+        
+        if(adminService.checkTokenValidity(header)){
+             commonMethodService.addCategory(category);
         return "Successfully Added";
+        }
+        return "Wrong token";
+       
     }
 
     @PutMapping("/category/update")
-    private String updateCategory(@RequestParam Long id, @RequestBody categories category){
-        return commonMethodService.updateCategory(id,category);
+    private String updateCategory(@RequestParam Long id, @RequestBody categories category , @RequestHeader String header) throws Exception{
+        
+        if(adminService.checkTokenValidity(header)){
+               return commonMethodService.updateCategory(id,category);
+
+        }
+        return "Wrong token";
     }
 
     @DeleteMapping("/category/delete")
-    private String deleteCategory(@RequestParam Long id){
+    private String deleteCategory(@RequestParam Long id , @RequestHeader String header) throws Exception{
+        
+        if(adminService.checkTokenValidity(header)){
         return commonMethodService.deleteCategory(id);
+        }
+        return "Wrong token";
     }
 
     /** Producer **/
 
     @GetMapping("/getproducers")
-    private List<producers> getAllProducers(){
+    private List<producers> getAllProducers( @RequestHeader String header) throws Exception{
+        if(adminService.checkTokenValidity(header)){
         return commonMethodService.getAllProducers();
+        }
+        return null;
     }
 
     @PostMapping("/producer/add")
-    private String addProducer(@RequestBody producers producer){
-        commonMethodService.addProducer(producer);
-        return "Successfully Added";
+    private String addProducer(@RequestBody producers producer ,@RequestHeader String header) throws Exception{
+        
+        if(adminService.checkTokenValidity(header)){
+commonMethodService.addProducer(producer);
+        return "Successfully Added";      
+        }
+        return "Wrong token";
+        
     }
 
     @PutMapping("/producer/update")
-    private String updateProducer(@RequestParam Long id, @RequestBody producers producer){
+    private String updateProducer(@RequestParam Long id, @RequestBody producers producer ,@RequestHeader String header) throws Exception{
+        if(adminService.checkTokenValidity(header)){
         return commonMethodService.updateProducer(id,producer);
+        }
+        return "Wrong token";
     }
 
     @DeleteMapping("/producer/delete")
-    private String deleteProducer(@RequestParam Long id){
+    private String deleteProducer(@RequestParam Long id , @RequestHeader String header)  throws Exception{
+        if(adminService.checkTokenValidity(header)){
         return commonMethodService.deleteProducer(id);
+        }
+        return "Wrong token";
     }
 
     /** StateCodes **/
     @GetMapping("/getstates")
-    private List<statecodes> getAllStates(){
+    private List<statecodes> getAllStates(@RequestHeader String header) throws Exception{
+        if(adminService.checkTokenValidity(header)){
         return commonMethodService.getAllStates();
+        }
+        return null;
     }
     @PostMapping("/states/add")
-    private String addStates(@RequestBody statecodes statecode){
-        commonMethodService.addStateCode(statecode);
+    private String addStates(@RequestBody statecodes statecode ,@RequestHeader String header) throws Exception{
+        if(adminService.checkTokenValidity(header)){
+commonMethodService.addStateCode(statecode);
         return "Successfully Added";
+        }
+        return "Wrong token";
+        
     }
 
     @PutMapping("/states/update")
-    private String updateStates(@RequestParam Long id, @RequestBody statecodes statecode){
+    private String updateStates(@RequestParam Long id, @RequestBody statecodes statecode ,@RequestHeader String header) throws Exception{
+        
+        if(adminService.checkTokenValidity(header)){
         return commonMethodService.updateStateCode(id,statecode);
+        }
+        return "Wrong token";
     }
 
     @DeleteMapping("/states/delete")
-    private String deleteStates(@RequestParam Long id){
+    private String deleteStates(@RequestParam Long id , @RequestHeader String header) throws Exception{
+        
+        if(adminService.checkTokenValidity(header)){
         return commonMethodService.deleteStateCode(id);
+        }
+        return "Wrong token";
     }
 
     /** Uncommon **/
 
     @PostMapping("/uncommon/add")
-    private Long addUncommon(@RequestBody uncommonProductDto uncommonProduct){
+    private Long addUncommon(@RequestBody uncommonProductDto uncommonProduct ,@RequestHeader String header) throws Exception{
+        
+        if(adminService.checkTokenValidity(header)){
         return commonMethodService.addUncommonProduction(uncommonProduct);
+        }
+        return null;
         //return "Successfully Added";
     }
 
     @PutMapping("/uncommon/update")
-    private String updateUncommon(@RequestParam Long id, @RequestBody uncommonProduct uncommonProduct){
+    private String updateUncommon(@RequestParam Long id, @RequestBody uncommonProduct uncommonProduct , @RequestHeader String header) throws Exception{
+        
+        if(adminService.checkTokenValidity(header)){
         return commonMethodService.updateUncommonProduction(id,uncommonProduct);
+        }
+        return "Wrong token";
     }
 
     @DeleteMapping("/uncommon/delete")
-    private String deleteUncommon(@RequestParam Long id){
+    private String deleteUncommon(@RequestParam Long id , @RequestHeader String header) throws Exception{
+        
+        if(adminService.checkTokenValidity(header)){
         return commonMethodService.deleteUncommonProduction(id);
+        }
+        return "Wrong token";
     }
 
     /** User **/
